Show Choice type values in prompt
