+++ .gitignore
@@ -105,4 +105,8 @@ com_crashlytics_export_strings.xml
 crashlytics.properties
 crashlytics-build.properties
 
+# virtualenv
 .venv/
+
+# vscode
+.vscode/
